Draft multipid daemon management.

+++ generic/process/daemon.sh
@@ -1,12 +1,42 @@
-#True if daemon is running.
-_daemonStatus() {
-	if [[ -e "$daemonPidFile" ]]
-	then
-		export daemonPID=$(cat "$daemonPidFile")
-	fi
+#Daemon signal handler is provided to allow entire process groups *not* to be killed, if this is not desired.
+#https://stackoverflow.com/questions/34438189/bash-sleep-process-not-getting-killed
+_daemonSendTERM() {
+	#pkill -TERM -P "$1"
 	
-	ps -p "$daemonPID" >/dev/null 2>&1 && return 0
+	kill -TERM "$1"
+}
+
+_daemonSendKILL() {
+	#pkill -KILL -P "$1"
+	
+	kill -KILL "$1"
+}
+
+#True if any daemon registered process is running.
+_daemonAction() {
+	[[ ! -e "$daemonPidFile" ]] && return 1
+	
+	local currentLine
+	local processedLine
+	local daemonStatus
+	
+	while IFS='' read -r currentLine || [[ -n "$currentLine" ]]; do
+		processedLine=$(echo "$currentLine" | tr -dc '0-9')
+		if [[ "$processedLine" != "" ]] && ps -p "$processedLine" >/dev/null 2>&1
+		then
+			daemonStatus="up"
+			[[ "$1" == "status" ]] && return 0
+			[[ "$1" == "terminate" ]] && _daemonSendTERM "$processedLine"
+			[[ "$1" == "kill" ]] && _daemonSendKILL "$processedLine"
+		fi
+	done < "$daemonPidFile"
+	
+	[[ "$daemonStatus" == "up" ]] && return 0
 	return 1
+}
+
+_daemonStatus() {
+	_daemonAction "status"
 }
 
 _waitForTermination() {
@@ -19,19 +49,52 @@ alias _waitForDaemon=_waitForTermination
 
 #Kills background process using PID file.
 _killDaemon() {
-	_daemonStatus && kill -TERM "$daemonPID" >/dev/null 2>&1
+	_daemonAction "terminate"
 	
 	_waitForTermination
 	
-	_daemonStatus && kill -KILL "$daemonPID" >/dev/null 2>&1
+	_daemonAction "kill"
 	
 	_waitForTermination
 	
-	rm -f "$daemonPidFile" >/dev/null 2>&1
+	#Do NOT detele this file unless daemon can be confirmed down.
+	! _daemonStatus && rm -f "$daemonPidFile" >/dev/null 2>&1
+}
+
+_cmdDaemon() {
+	_daemonStatus && return 1
+	
+	export isDaemon="true"
+	
+	echo "$$" >> "$daemonPidFile"
+	
+	"$@" &
+	
+	#Any PID which may be part of a daemon may be appended to this file.
+	echo "$!" >> "$daemonPidFile"
 }
 
 #Executes self in background (ie. as daemon).
 _execDaemon() {
-	"$scriptAbsoluteLocation" >/dev/null 2>&1 &
-	echo "$!" > "$daemonPidFile"
+	_daemonStatus && return 1
+	export isDaemon="true"
+	echo "$$" >> "$daemonPidFile"
+	_cmdDaemon "$scriptAbsoluteLocation"
+}
+
+_launchDaemon() {
+	_start
+	
+	_killDaemon
+	
+	
+	_execDaemon
+	while _daemonStatus
+	do
+		sleep 5
+	done
+	
+	
+	
+	_stop
 }
